Unit test updates.
