+++ functions/package.json
@@ -6,7 +6,7 @@
     "node": "8"
   },
   "scripts": {
-    "test": "npm run build && mocha -R ./test/unit/**",
+    "test": "npm run build && mocha ./test/unit/**",
     "test:cov": "istanbul cover .$(npm bin)/_mocha ./test/unit/**",
     "lint": "eslint -c .eslintrc ./src",
     "build": "npm run clean && babel src -d --out-dir dist",
@@ -24,8 +24,8 @@
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
     "express": "^4.16.3",
-    "firebase-admin": "^5.12.1",
-    "firebase-functions": "^1.0.4",
+    "firebase-admin": "^5.13.1",
+    "firebase-functions": "^2.0.0",
     "fs-extra": "^4.0.2",
     "glob": "^7.1.2",
     "googleapis": "^23.0.0",
@@ -44,9 +44,9 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "chai": "^4.1.2",
-    "firebase-functions-test": "^0.1.2",
+    "firebase-functions-test": "^0.1.3",
     "istanbul": "1.1.0-alpha.1",
-    "mocha": "^5.0.4",
-    "sinon": "^4.1.2"
+    "mocha": "^5.2.0",
+    "sinon": "^6.1.4"
   }
 }
