+++ functions/test/unit/index.spec.js
@@ -1,13 +1,11 @@
-import * as admin from 'firebase-admin' // eslint-disable-line no-unused-vars
-
 describe('Cloud Functions', () => {
   let myFunctions
-  let configStub
   let adminInitStub
   let admin
 
   before(() => {
     /* eslint-disable global-require */
+    admin = require('firebase-admin')
     adminInitStub = sinon.stub(admin, 'initializeApp')
     myFunctions = require(`${__dirname}/../../index`)
     /* eslint-enable global-require */
@@ -15,7 +13,6 @@ describe('Cloud Functions', () => {
 
   after(() => {
     // Restoring our stubs to the original methods.
-    configStub.restore()
     adminInitStub.restore()
   })
